Crate shams_time

Islamic prayer times and other time events that occur throughout a given day.

Example

    use shams_time::ShamsTimeSpec;
     
    // Specify location and calculation details
    let times = ShamsTimeSpec::new().method(6)
                                   .coordinates(29.366667, 47.966667)
                                   .timezone(3.0).temperature(25.0)
                                   .pressure(1010.0).altitude(20.0)
                                   .today().calculate();
     
    // Get the midpoint between sunset and sunrise
    let (hours, minutes, _) = times.midnight().expect("unsupported location");

    // Get Al-fajr prayer time
    let (fajr_hours, _, _, _) = times.fajr();
     
    println!("True midnight at {}:{:02}; {} hours after fajr", 
        hours, minutes, hours - fajr_hours);

Copyright

Copyright © 2015 Josh Morin JoshMorin@gmx.com

This program is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.

You should have received a copy of the GNU General Public License along with this program. If not, see http://www.gnu.org/licenses/.

Structs

ShamsTime
A Collection of times of interest

ShamsTimeSpec
Environment and Calculation Specification.