Struct shams_time::ShamsTime

pub struct ShamsTime {
    // some fields omitted
}

A Collection of times of interest

Methods

impl ShamsTime

fn date(&self) -> Date

fn fajr(&self) -> (Option<i64>, bool)

Time of fajr prayer.

fn shurooq(&self) -> (Option<i64>, bool)

Time of Shurooq.

fn zuhr(&self) -> (Option<i64>, bool)

Time of Duhur prayer.

fn asr(&self) -> (Option<i64>, bool)

Time of Aser prayer.

fn maghrib(&self) -> (Option<i64>, bool)

Time of Magrib prayer.

fn isha(&self) -> (Option<i64>, bool)

Time of Isha prayer.

fn sunrise(&self) -> Option<i64>

Time of Sunrise. Will return None if ITL can not reliably calculate the necessary values.

fn sunset(&self) -> Option<i64>

Time of Sunset. Will return None if ITL can not reliably calculate the necessary values.

fn daytime(&self) -> Option<i64>

Durration of daytime (the time between sunrise and sunset)

fn nighttime(&self) -> Option<i64>

Durration of nighttime (the time between sunset and sunrise)

Currently, a slight error will occure due to the fact that sunrise time is taken from the current date rather than the next.

fn noon(&self) -> Option<i64>

Returns the midpoint between sunrise and sunset. Will return None if ITL can not reliably calculate the necessary values.

fn midnight(&self) -> Option<i64>

Returns the midpoint between sunset and sunrise. Will return None if ITL can not reliably calculate the necessary values.

Since this function internally uses the sunrise of the current day rather than the day after, the result will be slightly inaccurate.

Note that hours can go over 24, which indicates that midnight occurs the following day.

Note that this value does not represent the end of the Isha time if it ends at "midnight". In Islam, daytime starts at dusk, rather than sunrise.