Struct shams_time::ShamsTimeSpec

pub struct ShamsTimeSpec {
    // some fields omitted
}

Environment and Calculation Specification.

This structure specifies the necessary details for generating correct times of interest.

Environment Details

Latitude and Longitude

Elevation

Atmospheric Pressure

Temperature

Time Details

Prayer time Calculation Methods

Timezone

Date

DST state

Extended Method Details

Unless this structure is populated with all the necessary information calculations require, calculations will yield incorrect results. The new function does not initialize the specification to a usable state. Setting information is done by call the appropriate method.

Extended Details

Atmospheric pressure varies with elevation, location, temperature, and time of year. For the eastern coast of Arabia, pressure can range from <1000 millibars up to around 1020. The standard value is 1010.0.

Elevation, temperature, location coordinates, Time zone, and DST can all be determined with a quick web search.

Which method to use varies from one region to another. If you don't care about the correctness of prayer times, you can set it to 0.

Methods

impl ShamsTimeSpec

fn new() -> ShamsTimeSpec

Empty ShamsTimeSpec Object. This object must be converted to a ShamsTime object by calling calculate or year before it can be used to infer times.

Converting this object without specifying all the required information will yield incorrect results.

fn city(name: &str) -> Option<ShamsTimeSpec>

Attempts to set method, latitude, longitude, elevation, pressure and timezone. This deviates from the builder pattern because only a small set of cities have these values defined.

City names may be supplied in Arabic or English. If multiple cities with the same name exist, include the name of the country. For Example, the strings "Al-Quds", "Al-Quds, Palestine", "القدس", and "القدس، فلسطين" are all equivalent and will set the same value.

Current cities defined are:

Makkah, Saudi Arabia

Medina, Saudi Arabia

Al-Quds, Palestine

Jakarta, Indonesia

Bandar Seri Begawan, Brunei

Kuwait, Kuwait

If your city is not included in this list, please consider taking the time to submit the necessary information to the bug tracker. Thank you.

fn method(&mut self, id: u8) -> &mut ShamsTimeSpec

Sets the method of calculating prayer times.

Methods can be one of the following:

       0: none. Set to default or 0
       1: Egyptian General Authority of Survey
       2: University of Islamic Sciences, Karachi (Shaf'i)
       3: University of Islamic Sciences, Karachi (Hanafi)
       4: Islamic Society of North America
       5: Muslim World League (MWL)
       6: Umm Al-Qurra, Saudi Arabia
       7: Fixed Ishaa Interval (always 90)
       8: Egyptian General Authority of Survey (Egypt)
       9: Umm Al-Qurra Ramadan, Saudi Arabia
      10: Moonsighting Committee Worldwide
      11: Morocco Awqaf, Morocco

fn coordinates(&mut self, lat: f64, lon: f64) -> &mut ShamsTimeSpec

Sets latitude and longitude locations.

fn timezone(&mut self, offset: f64) -> &mut ShamsTimeSpec

Sets the GMT/UTC offset to offset.

All passed to this object or resulting objects should be relative to this offset.

fn temperature(&mut self, celsius: f64) -> &mut ShamsTimeSpec

Sets the temperature to celsius. Temperature values are in Celsius.

fn pressure(&mut self, millibar: f64) -> &mut ShamsTimeSpec

Sets the atmospheric pressure to millibar. Pressure values are in millibars.

fn elevation(&mut self, meters: f64) -> &mut ShamsTimeSpec

Sets the elevation to meters meters above sea level.

fn dst(&mut self, enable: bool) -> &mut ShamsTimeSpec

Sets the state of daylight saving time.

fn today(&mut self) -> &mut ShamsTimeSpec

Sets date of interest to today.

fn date(&mut self, year: u16, month: u16, day: u16) -> &mut ShamsTimeSpec

Sets the date of interest to the date represented by its arguments.

Ranges are inclusive: - day range is [1, 31] - month range: [1, 12] - year is the common 4 digit representation.

fn unix_date(&mut self, ts: i64) -> &mut ShamsTimeSpec

Sets the date to the date defined by a time stamp.

fn fajr_angle(&mut self, degrees: f64) -> &mut ShamsTimeSpec

Sets the Fajr angle in degrees.

This overrides the value defined by a previous call to method.

fn isha_angle(&mut self, degrees: f64) -> &mut ShamsTimeSpec

Sets the Isha angle in degrees.

This overrides the value defined by a previous call to method.

fn fajr_interval(&mut self, minutes: i32) -> &mut ShamsTimeSpec

Sets the Fajr interval in minutes between Fajr and Shurooq.

This overrides the value defined by a previous call to method.

fn isha_interval(&mut self, minutes: i32) -> &mut ShamsTimeSpec

Sets the Isha interval in minutes between Magrib and Isha

This overrides the value defined by a previous call to method.

fn aser_mathhab(&mut self, id: i32) -> &mut ShamsTimeSpec

Selects the Asr time calculation method; 1: Shafi, 2: Hanafi.

This overrides the value defined by a previous call to method.

fn nearest_latitude(&mut self, angle: f64) -> &mut ShamsTimeSpec

(used for extreme methods: refer to ITL's documentation)

fn extreme(&mut self, extreme: i32) -> &mut ShamsTimeSpec

(used for extreme methods: refer to ITL's documentation)

fn calculate(&self) -> Result<ShamsTime, String>

Calculates the times of interest based on the supplied information. The specification object can be reused post calculation.

fn year(&self, year: u16) -> Result<YearIter, String>

Returns an iterator that produces a ShamsTime for each day of the specified year.