Crate shams_time

Islamic prayer times and other time events that occur throughout a given day.

Examples: The time between Al-Fajr and true midnight

extern crate shams_time;
extern crate time;
 
fn main() {
    use shams_time::ShamsTimeSpec;
    use time::Duration;

    // Specify location and calculation details
    let st = ShamsTimeSpec::new().method(6)
                                 .coordinates(29.366667, 47.966667)
                                 .timezone(3.0).temperature(25.0)
                                 .pressure(1010.0).altitude(20.0)
                                 .today().calculate();
     
    // Get true midnight and Al-fajr
    let midnight = st.midnight().expect("midnight sun?");
    let fajr = st.fajr().0.expect("unsupported location");
     
    println!("True midnight is {} hours after fajr", 
        Duration::seconds(midnight - fajr).num_hours());

}

Example: Longest day of the year

extern crate shams_time;
extern crate time;
 
fn main() {
    use shams_time::ShamsTimeSpec;
    use time::Duration;
    use std::default::Default;
     
    // Get an iterator that yeilds a ShamsTime object for every day of the 
    // year 2003
    let year = ShamsTimeSpec::city("Makkah").expect("unknown city").year(2003);
     
    // iterate of all days and search for longest day
    let (date, durration) = year.fold((Default::default(), 0), |max, day| {
        let daytime = day.daytime().expect("no sunset or sunrise");
        if daytime > max.1 {
            (day.date(), daytime)
        } 
        else {
            max
        }
    });

    // converting unix time to hours, minutes, and seconds
    let durration = Duration::seconds(durration);
    println!(concat!("longest day in Makkah for the year {} was on {}/{}, ", 
                     "lasting for {} hours."), 
                      date.year, date.day, date.month,
                      durration.num_hours());
}
 

Copyright

Copyright © 2015 Josh Morin JoshMorin@gmx.com

This program is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.

You should have received a copy of the GNU General Public License along with this program. If not, see http://www.gnu.org/licenses/.

Structs

Date

ShamsTime
A Collection of times of interest

ShamsTimeSpec
Environment and Calculation Specification.

YearIter
Iterates over all days of a given year.