Struct shams_time::ShamsTime

pub struct ShamsTime {
    // some fields omitted
}

A Collection of times of interest for a specific date

Some methods will use Extreme Methods to preform the required calculates if configured to do so. See the Extreme Methods section in the top level of the documentation for a description of what extreme methods are and what they entail.

Methods

impl ShamsTime

fn date(&self) -> Date

Returns the date for which the results in this object apply.

fn fajr(&self) -> (Option<Time>, bool)

Time of Fajr prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn next_fajr(&self) -> (Option<Time>, bool)

Time of next-day Fajr prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn shurooq(&self) -> (Option<Time>, bool)

Time of Shurooq prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn next_shurooq(&self) -> (Option<Time>, bool)

Time of next-day Shurooq prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn zuhr(&self) -> (Option<Time>, bool)

Time of Zuhr prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn asr(&self) -> (Option<Time>, bool)

Time of Asr prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn maghrib(&self) -> (Option<Time>, bool)

Time of Maghrib prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn isha(&self) -> (Option<Time>, bool)

Time of Isha prayer if calculable and a flag to signal if the time was calculated using extreme methods.

fn sunrise(&self) -> Option<Time>

Time of Sunrise.

Returns None if this value can not be reliably calculated.

fn next_sunrise(&self) -> Option<Time>

Time of next-day sunrise.

Returns None if this value can not be reliably calculated.

fn sunset(&self) -> Option<Time>

Time of Sunset.

Returns None if this value can not be reliably calculated.

fn daytime(&self) -> Option<Duration>

Duration (no leap seconds) of daytime in seconds: The time between sunrise and sunset.

Returns None if this value can not be reliably calculated.

fn nighttime(&self) -> Option<Duration>

Dur ration (no leap seconds) of nighttime in seconds: The time between sunset and sunrise.

Returns None if this value can not be reliably calculated.

fn noon(&self) -> Option<Time>

Returns POSIX time of midday; between sunrise and sunset.

Returns None if this value can not be reliably calculated.

fn midnight(&self) -> Option<Time>

Returns the midpoint between sunset and sunrise.

Returns None if this value can not be reliably calculated.

Note that this value does not represent the end of the Isha time if it ends at "midnight". In Islam, daytime starts at dusk, rather than sunrise.

fn angle_time(&self, angle: f64) -> Option<Time>

Returns a time at which the sun would be perceived to be at angle with the location from which it rose.

The times this function returns correspond to what humans would perceive as the reasonable values.

To maintain perception accuracy, angle does not linearly correlate with time or relative location of the sun. See phase_time.

Also, the time at angle 0.0 does not necessarily equate to that of 60.0. Angles 0.0 and 360.0 correspond to the sunrises of the current day and the day the follows it, respectively.

Angle values must be between 0.0 and 360.0. if this is not the case, angle will be clipped to those boundaries.

fn phase_time(&self, phase: f64) -> Option<Time>

Returns the POSIX time when the sun is at a phase defined in the table bellow:

Normalized phase of 0.0 implies sunrise.

Normalized phase of 0.5 implies noon.

Normalized phase of 1.0 implies sunset.

Normalized phase of 1.5 implies midnight.

Normalized phase of 2.0 implies the sunrise of the next day

Phase value must be between 0.0 and 2.0, inclusive. If this isn't the case, phase will be clipped to the nearest boundary.

Note: The phase value does not linearly correlate with the sun's actual angle (time of day); the duration between 0.0 and 1.0 will likely be different than the duration between 1.0 and 2.0, depending on the location and time of year.

fn time_phase(&self, time: Time) -> Option<f64>